# perf: reduce python overhead for awkward backend