--- a/src/vector/_compute/lorentz/boostZ_beta.py
+++ b/src/vector/_compute/lorentz/boostZ_beta.py
@@ -218,7 +218,7 @@ def dispatch(beta: typing.Any, v: typing.Any) -> typing.Any:
     with numpy.errstate(all="ignore"):
         return v._wrap_result(
             _flavor_of(v),
-            function(
+            v._wrap_dispatched_function(function)(
                 v.lib,
                 beta,
                 *v.azimuthal.elements,