--- a/src/vector/_compute/lorentz/boostZ_gamma.py
+++ b/src/vector/_compute/lorentz/boostZ_gamma.py
@@ -218,7 +218,7 @@ def dispatch(gamma: typing.Any, v: typing.Any) -> typing.Any:
     with numpy.errstate(all="ignore"):
         return v._wrap_result(
             _flavor_of(v),
-            function(
+            v._wrap_dispatched_function(function)(
                 v.lib,
                 gamma,
                 *v.azimuthal.elements,