--- a/src/vector/_compute/lorentz/deltaRapidityPhi.py
+++ b/src/vector/_compute/lorentz/deltaRapidityPhi.py
@@ -110,9 +110,10 @@ def dispatch(
         ),
     )
     with numpy.errstate(all="ignore"):
-        return _handler_of(v1, v2)._wrap_result(
+        handler = _handler_of(v1, v2)
+        return handler._wrap_result(
             _flavor_of(v1, v2),
-            function(
+            handler._wrap_dispatched_function(function)(
                 _lib_of(v1, v2),
                 *v1.azimuthal.elements,
                 *v1.longitudinal.elements,