--- a/src/vector/_compute/lorentz/is_spacelike.py
+++ b/src/vector/_compute/lorentz/is_spacelike.py
@@ -66,7 +66,7 @@ def dispatch(tolerance: typing.Any, v: typing.Any) -> typing.Any:
     with numpy.errstate(all="ignore"):
         return v._wrap_result(
             _flavor_of(v),
-            function(
+            v._wrap_dispatched_function(function)(
                 v.lib,
                 tolerance,
                 *v.azimuthal.elements,