--- a/src/vector/_compute/lorentz/scale.py
+++ b/src/vector/_compute/lorentz/scale.py
@@ -181,7 +181,7 @@ def dispatch(factor: typing.Any, v: typing.Any) -> typing.Any:
     with numpy.errstate(all="ignore"):
         return v._wrap_result(
             _flavor_of(v),
-            function(
+            v._wrap_dispatched_function(function)(
                 v.lib,
                 factor,
                 *v.azimuthal.elements,