--- a/src/vector/_compute/lorentz/unit.py
+++ b/src/vector/_compute/lorentz/unit.py
@@ -248,7 +248,7 @@ def dispatch(v: typing.Any) -> typing.Any:
     with numpy.errstate(all="ignore"):
         return v._wrap_result(
             _flavor_of(v),
-            function(
+            v._wrap_dispatched_function(function)(
                 v.lib,
                 *v.azimuthal.elements,
                 *v.longitudinal.elements,