--- a/src/vector/_compute/planar/is_parallel.py
+++ b/src/vector/_compute/planar/is_parallel.py
@@ -57,9 +57,10 @@ def dispatch(tolerance: typing.Any, v1: typing.Any, v2: typing.Any) -> typing.An
         ),
     )
     with numpy.errstate(all="ignore"):
-        return _handler_of(v1, v2)._wrap_result(
+        handler = _handler_of(v1, v2)
+        return handler._wrap_result(
             _flavor_of(v1, v2),
-            function(
+            handler._wrap_dispatched_function(function)(
                 _lib_of(v1, v2),
                 tolerance,
                 *v1.azimuthal.elements,