--- a/src/vector/_compute/planar/isclose.py
+++ b/src/vector/_compute/planar/isclose.py
@@ -94,9 +94,10 @@ def dispatch(
         ),
     )
     with numpy.errstate(all="ignore"):
-        return _handler_of(v1, v2)._wrap_result(
+        handler = _handler_of(v1, v2)
+        return handler._wrap_result(
             _flavor_of(v1, v2),
-            function(
+            handler._wrap_dispatched_function(function)(
                 _lib_of(v1, v2),
                 rtol,
                 atol,