--- a/src/vector/_compute/planar/phi.py
+++ b/src/vector/_compute/planar/phi.py
@@ -33,6 +33,9 @@ def rhophi(lib, rho, phi):
     return phi
 
 
+rhophi.__awkward_transform_allowed__ = False  # type:ignore[attr-defined]
+
+
 dispatch_map = {
     (AzimuthalXY,): (xy, float),
     (AzimuthalRhoPhi,): (rhophi, float),