--- a/src/vector/_compute/planar/rotateZ.py
+++ b/src/vector/_compute/planar/rotateZ.py
@@ -49,7 +49,7 @@ def dispatch(angle: typing.Any, v: typing.Any) -> typing.Any:
     with numpy.errstate(all="ignore"):
         return v._wrap_result(
             _flavor_of(v),
-            function(v.lib, angle, *v.azimuthal.elements),
+            v._wrap_dispatched_function(function)(v.lib, angle, *v.azimuthal.elements),
             returns,
             1,
         )