--- a/src/vector/_compute/planar/scale.py
+++ b/src/vector/_compute/planar/scale.py
@@ -50,7 +50,7 @@ def dispatch(factor: typing.Any, v: typing.Any) -> typing.Any:
     with numpy.errstate(all="ignore"):
         return v._wrap_result(
             _flavor_of(v),
-            function(v.lib, factor, *v.azimuthal.elements),
+            v._wrap_dispatched_function(function)(v.lib, factor, *v.azimuthal.elements),
             returns,
             1,
         )