--- a/src/vector/_compute/planar/transform2D.py
+++ b/src/vector/_compute/planar/transform2D.py
@@ -50,7 +50,7 @@ def dispatch(obj: typing.Any, v: typing.Any) -> typing.Any:
     with numpy.errstate(all="ignore"):
         return v._wrap_result(
             _flavor_of(v),
-            function(
+            v._wrap_dispatched_function(function)(
                 v.lib, obj["xx"], obj["xy"], obj["yx"], obj["yy"], *v.azimuthal.elements
             ),
             returns,