--- a/src/vector/_compute/planar/y.py
+++ b/src/vector/_compute/planar/y.py
@@ -29,6 +29,9 @@ def xy(lib, x, y):
     return y
 
 
+xy.__awkward_transform_allowed__ = False  # type:ignore[attr-defined]
+
+
 def rhophi(lib, rho, phi):
     return rho * lib.sin(phi)