… src/vector/_compute/spatial/cottheta.py
… src/vector/_compute/spatial/cross.py
… src/vector/_compute/spatial/deltaR.py
… src/vector/_compute/spatial/deltaR2.py
… src/vector/_compute/spatial/deltaangle.py
… src/vector/_compute/spatial/deltaeta.py
… src/vector/_compute/spatial/dot.py
… src/vector/_compute/spatial/equal.py
… src/vector/_compute/spatial/eta.py
… src/vector/_compute/spatial/is_antiparallel.py
… src/vector/_compute/spatial/is_parallel.py
… src/vector/_compute/spatial/is_perpendicular.py
… src/vector/_compute/spatial/isclose.py
… src/vector/_compute/spatial/mag.py
… src/vector/_compute/spatial/mag2.py
… src/vector/_compute/spatial/not_equal.py
… src/vector/_compute/spatial/rotateX.py
… src/vector/_compute/spatial/rotateY.py
… src/vector/_compute/spatial/rotate_axis.py
… src/vector/_compute/spatial/rotate_euler.py
… src/vector/_compute/spatial/rotate_quaternion.py
… src/vector/_compute/spatial/scale.py
… src/vector/_compute/spatial/subtract.py
… src/vector/_compute/spatial/theta.py
… src/vector/_compute/spatial/transform3D.py
… src/vector/_compute/spatial/unit.py
… src/vector/_compute/spatial/z.py
… src/vector/_methods.py
… src/vector/backends/awkward.py
… src/vector/backends/numpy.py