@pfackeldey 2025-01-13
In context of the trijet mass calculation of the AGC (ref: https://github.com/scikit-hep/awkward/pull/3359), this brings the runtime further down to (from ~25ms):
```python
# 100k events (cpu backend)
In [1]: %timeit calculate_trijet_mass(events)
19.1 ms ± 203 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# typetracer
In [2]: %timeit calculate_trijet_mass(tt)
18.2 ms ± 49.6 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```