@ianna 2025-01-15
> Hi @ianna, thanks for your review! Do you have some opinions about the name of this function/decorator? And/or how "publicly" exposed it should be? I was wondering also, if we should add a "Tipp" to the awkward-array documentation that if one wants to apply a batch of e.g. ufuncs, it might be beneficial to do this in a single layout traversal with `ak.transform`. Is this somewhere already, if not, do you think it's worth to add somewhere?

It's a good idea to add a performance hint in the documentation. Perhaps, even a new section with "good practices" or "performance guide"?

Do you mean something like:

```python
@broadcast_once(axis=0)
def my_func(arr) 
...
```