@ianna 2025-01-15
> > > Hi @ianna, thanks for your review! Do you have some opinions about the name of this function/decorator? And/or how "publicly" exposed it should be? I was wondering also, if we should add a "Tipp" to the awkward-array documentation that if one wants to apply a batch of e.g. ufuncs, it might be beneficial to do this in a single layout traversal with `ak.transform`. Is this somewhere already, if not, do you think it's worth to add somewhere?
> > 
> > 
> > It's a good idea to add a performance hint in the documentation. Perhaps, even a new section with "good practices" or "performance guide"?
> > Do you mean something like:
> > ```python
> > @broadcast_once(axis=0)
> > def my_func(arr) 
> > ...
> > ```
> 
> `ak.transform` is already doing the layout traversal (e.g. in context of broadcasting) only once. So we don't need to add a new function. It would just be a hint in the `ak.transform` documentation. I don't think we can really specify an axis, because all inputs to a transform need to be broadcasted together ([numpy's broadcasting also doesn't support specifying an axis](https://numpy.org/doc/2.1/reference/generated/numpy.broadcast_arrays.html#numpy.broadcast_arrays)).
> 
> Ok, I'd go ahead and think about how to add a tipp/hint in awkward's documentation for `ak.transform` 👍.

if you are done with the PR, I can merge it :-)